% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{raw_data}
\alias{raw_data}
\title{Example Dataset: An Example of Raw Data}
\format{
A list containing:
\describe{
\item{id}{Individual ID}
\item{time}{the time points}
\item{space}{the number of perches where bird locations are captured}
\item{x}{the x-coordinate of location information}
\item{y}{the y-coordinate of location information}
\item{Replicate}{Replicate 1-4}
}
}
\source{
Generated from XYZ
}
\usage{
raw_data
}
\description{
This dataset contains spatial information of zebra finches in Replicate 1
}
\keyword{datasets}
