% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvf.R
\name{sdm}
\alias{sdm}
\title{Returns the sum of squared deviations from the mean}
\usage{
sdm(data)
}
\arguments{
\item{data}{A numeric vector}
}
\value{
The sum of squared deviations from the mean
}
\description{
Returns the sum of squared deviations from the mean
}
\examples{
sdm(c(1, 2, 3))
}
