% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zf_ind.R
\docType{data}
\name{zf_ind}
\alias{zf_ind}
\title{Example Dataset: zf_ind}
\format{
A list containing:
\describe{
\item{Ind_ID}{Individual ID}
\item{Experiment}{Experiment/Round 1 or 2}
\item{Room}{Room number 7 or 8}
\item{sex}{Sex of the individual}
\item{QR}{the number of the QR code on the birds}
\item{Replicate}{Replicate 1-4}
}
}
\source{
Generated from XYZ
}
\usage{
zf_ind
}
\description{
This dataset contains metadata for the experiments.
}
\keyword{datasets}
