% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zf_rep1_null2.R
\docType{data}
\name{zf_rep1_null2}
\alias{zf_rep1_null2}
\title{Example Dataset: zf_rep1_null2}
\format{
An array containing:
\describe{
\item{First/Second dimension}{Individual ID in replicate 1}
\item{Third dimension}{Days}
}
}
\source{
Generated from XYZ
}
\usage{
zf_rep1_null2
}
\description{
This dataset contains the clumping social networks in null model 2 of zebra finches in Replicate 1 during 30 days.
}
\keyword{datasets}
