% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zf_rep3_real.R
\docType{data}
\name{zf_rep3_real}
\alias{zf_rep3_real}
\title{Example Dataset: zf_rep3_real}
\format{
An array containing:
\describe{
\item{First/Second dimension}{Individual ID in replicate 3}
\item{Third dimension}{Days}
}
}
\source{
Generated from XYZ
}
\usage{
zf_rep3_real
}
\description{
This dataset contains the clumping social networks of
zebra finches in Replicate 3 during 30 days.
}
\keyword{datasets}
